+++ gambatte/libgambatte/src/sound/length_counter.cpp
@@ -22,7 +22,7 @@
 
 namespace gambatte {
 
-LengthCounter::LengthCounter(MasterDisabler &disabler, const unsigned mask) :
+LengthCounter::LengthCounter(MasterDisabler &disabler, const uint32_t mask) :
 	disableMaster(disabler),
 	lengthMask(mask)
 {
@@ -35,17 +35,17 @@ void LengthCounter::event() {
 	disableMaster();
 }
 
-void LengthCounter::nr1Change(const unsigned newNr1, const unsigned nr4, const unsigned long cycleCounter) {
+void LengthCounter::nr1Change(const uint32_t newNr1, const uint32_t nr4, const uint32_t cycleCounter) {
 	lengthCounter = (~newNr1 & lengthMask) + 1;
-	counter = (nr4 & 0x40) ?( (cycleCounter >> 13) + lengthCounter) << 13 : static_cast<unsigned long>(COUNTER_DISABLED);
+	counter = (nr4 & 0x40) ?( (cycleCounter >> 13) + lengthCounter) << 13 : static_cast<uint32_t>(COUNTER_DISABLED);
 }
 
-void LengthCounter::nr4Change(const unsigned oldNr4, const unsigned newNr4, const unsigned long cycleCounter) {
+void LengthCounter::nr4Change(const uint32_t oldNr4, const uint32_t newNr4, const uint32_t cycleCounter) {
 	if (counter != COUNTER_DISABLED)
 		lengthCounter = (counter >> 13) - (cycleCounter >> 13);
 	
 	{
-		unsigned dec = 0;
+		uint32_t dec = 0;
 		
 		if (newNr4 & 0x40) {
 			dec = ~cycleCounter >> 12 & 1;
@@ -66,7 +66,7 @@ void LengthCounter::nr4Change(const unsigned oldNr4, const unsigned newNr4, cons
 		counter = COUNTER_DISABLED;
 }
 
-void LengthCounter::loadState(const SaveState::SPU::LCounter &lstate, const unsigned long cc) {
+void LengthCounter::loadState(const SaveState::SPU::LCounter &lstate, const uint32_t cc) {
 	counter = std::max(lstate.counter, cc);
 	lengthCounter = lstate.lengthCounter;
 }
